fix: AuthRegisterForm 이름변경, 유효성 검사 수정
